FDK-817 refactored glue code, removed redundant code
