+++ applications/end2end-test/src/test/java/no/dcat/bddtest/cucumber/glue/BackgroundPage.java
@@ -3,11 +3,14 @@ package no.dcat.bddtest.cucumber.glue;
 import cucumber.api.java.After;
 import cucumber.api.java.Before;
 import cucumber.api.java.en.Given;
+
 import no.dcat.bddtest.cucumber.model.ThemeCountSmall;
 import no.dcat.bddtest.elasticsearch.client.DeleteIndex;
 import no.dcat.harvester.crawler.Loader;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.core.io.Resource;
 import org.springframework.web.client.RestTemplate;
@@ -29,8 +32,8 @@ public class BackgroundPage extends CommonPage {
     private static Logger logger = LoggerFactory.getLogger(BackgroundPage.class);
     private final String index = "dcat";
 
-    private final String portalHostname = "localhost"; // getEnv("fdk.hostname");
-    private int portalPort = 8080; //getEnvInt("fdk.port");
+    private final String portalHostname = "localhost";
+    private int portalPort = 8080;
 
     @Before
     public void setup() {
@@ -46,25 +49,26 @@ public class BackgroundPage extends CommonPage {
     public void cleanElasticSearch() throws Throwable {
         String hostname = "localhost";
         int port = 9300;
-        //String hostname = getEnv("elasticsearch.hostname");
-        //int port = getEnvInt("elasticsearch.port");
 
         new DeleteIndex(hostname, port).deleteIndex(index);
     }
+
 
     @Given("^I load the \"([^\"]*)\" dataset\\.$")
     public void loadDataset(String filename) throws IOException {
         deleteAndLoad(filename);
     }
 
-    @Given("^Elasticsearch kjører")
+
+    @Given("^Elasticsearch is running")
     public void elasticSearchIsRunning() {
         RestTemplate restTemplate = new RestTemplate();
         String health = restTemplate.getForObject("http://localhost:9200/_cluster/health", String.class);
         assertThat(health, is(not(nullValue())));
     }
 
-    @Given("^bruker datasett (.*).ttl")
+
+    @Given("^uses dataset (.*).ttl")
     public void setupTestData(String datasett) throws IOException {
 
         logger.info("setupTestData(datasett: '{}')", datasett);
@@ -79,6 +83,13 @@ public class BackgroundPage extends CommonPage {
         deleteLoadAndWait(datasett + ".ttl", waitFor);
 
     }
+
+
+    @Given("^Search portal is open in web browser")
+    public void openBrowserToHomepage() {
+        driver.get("http://" + portalHostname + ":" + portalPort + "/");
+    }
+
 
     private int getThemeCount() {
         RestTemplate restTemplate1 = new RestTemplate();
@@ -95,10 +106,11 @@ public class BackgroundPage extends CommonPage {
         await().atMost(30, SECONDS).until(waitFor);
     }
 
+
     private void deleteAndLoad(String datasett) throws IOException {
 
-        String hostname = "localhost"; //getEnv("elasticsearch.hostname");
-        int port = 9300; //getEnvInt("elasticsearch.port");
+        String hostname = "localhost";
+        int port = 9300;
 
         new DeleteIndex(hostname, port).deleteIndex(index);
 
@@ -113,17 +125,6 @@ public class BackgroundPage extends CommonPage {
         waitForHarvesterToComplete();
 
         refreshElasticsearch(hostname, port, "elasticsearch");
-
-
     }
 
-    @Given("^man har åpnet Fellesdatakatalog i en nettleser")
-    public void openBrowserToHomepage() {
-        driver.get("http://" + portalHostname + ":" + portalPort + "/");
-    }
-
-
-    protected String getPage() {
-        return null;
-    }
 }
