+++ applications/end2end-test/src/test/java/no/dcat/bddtest/cucumber/glue/CommonPage.java
@@ -18,8 +18,7 @@ import org.openqa.selenium.support.ui.WebDriverWait;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.client.RestTemplate;
-
-;import java.util.Map;
+import java.util.Map;
 
 import static org.junit.Assert.assertTrue;
 
@@ -29,8 +28,6 @@ import static org.junit.Assert.assertTrue;
 public abstract class CommonPage extends SpringIntegrationTestConfigIT {
     private final Logger logger = LoggerFactory.getLogger(CommonPage.class);
     WebDriver driver = null;
-
-    public static final String PORTAL_URL = "http://localhost:8080"; // = "http://fdk-por-fellesdatakatalog-ut1.ose-npc.brreg.no/"; //"http://portal-fdk.tt1.brreg.no";
 
 
     public void openPage(String page) {
@@ -116,7 +113,6 @@ public abstract class CommonPage extends SpringIntegrationTestConfigIT {
 
     void waitForHarvesterToComplete(){
 
-
         int maxTries = 60;
         while(true){
             logger.info("Waiting for harvester to become idle.");
@@ -144,6 +140,7 @@ public abstract class CommonPage extends SpringIntegrationTestConfigIT {
         }
 
     }
+
 
     private boolean harvesterIsIdle() {
         Map<String, Boolean> forObject = new RestTemplate().getForObject("http://localhost:8081/api/admin/isIdle", Map.class);
