+++ applications/end2end-test/src/test/resources/feature/HarvestedDatasetsVisibleInPortal.feature
@@ -2,7 +2,7 @@ Feature: Harvested datasets become visible in portal
 
   Scenario: Datasets from harvested catalog is available in search portal
     Given I start with empty elasticsearch index.
-    When I open the admin portal
+    When I open the harvester administration page
     And I select harvest "difi" catalog
     Then the following dataset detail pages shall exist in search portal:
       | http://data.norge.no/node/581 |
